Move the setting of FQDN/PublicHostname in private cluster to the fakerp
also don't make a dns record for it, as it's not used.
