--- pkg/fakerp/codecov_dummy_test.go
@@ -1,12 +1,0 @@
-package fakerp
-
-import (
-	"testing"
-)
-
-// This file exists because this package has no unit tests.  Codecov does not
-// report packages with no unit tests as 0% coverage, incorrectly inflating our
-// coverage statistics.  When unit tests are added to this package, this file
-// can be removed.
-
-func TestDummyCodeCov(t *testing.T) {}
